+++ packages/xrpl/snippets/src/getTransaction.ts
@@ -19,10 +19,11 @@ async function getTransaction(): Promise<void> {
     })
     console.log(tx)
 
-    // The meta field would be a string(hex) when the `binary` parameter is `true` for the `tx` request.
+    // The meta field can be undefined if the transaction has not been validated yet (and so has not changed the ledger).
     if (tx.result.meta == null) {
       throw new Error('meta not included in the response')
     }
+
     /*
      * delivered_amount is the amount actually received by the destination account.
      * Use this field to determine how much was delivered, regardless of whether the transaction is a partial payment.