+++ packages/xrpl/src/utils/getNFTokenID.ts
@@ -1,4 +1,5 @@
 import flatMap from 'lodash/flatMap'
+import { decode } from 'ripple-binary-codec'
 
 import {
   CreatedNode,
@@ -16,22 +17,39 @@ interface NFToken {
 }
 
 /**
+ * Ensures that the metadata is in a deserialized format to parse.
+ *
+ * @param meta - the metadata from a `tx` method call. Can be in json format or binary format.
+ * @returns the metadata in a deserialized format.
+ */
+function ensureDecodedMeta(
+  meta: TransactionMetadata | string,
+): TransactionMetadata {
+  if (typeof meta === 'string') {
+    // eslint-disable-next-line @typescript-eslint/consistent-type-assertions -- Meta is either metadata or serialized metadata.
+    return decode(meta) as unknown as TransactionMetadata
+  }
+  return meta
+}
+
+/**
  * Gets the NFTokenID for an NFT recently minted with NFTokenMint.
  *
- * @param meta - Metadata from the response to submitting an NFTokenMint transaction.
+ * @param meta - Metadata from the response to submitting and waiting for an NFTokenMint transaction or from a `tx` method call.
  * @returns The NFTokenID for the minted NFT.
  * @throws if meta is not TransactionMetadata.
  */
 export default function getNFTokenID(
-  meta: TransactionMetadata,
+  meta: TransactionMetadata | string | undefined,
 ): string | undefined {
-  /* eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- Provides a nicer error for js users */
-  if (meta.AffectedNodes === undefined) {
-    throw new TypeError(`Unable to parse the parameter given to getNFTokenID. 
-    'meta' must be the metadata from an NFTokenMint transaction. Received ${JSON.stringify(
-      meta,
-    )} instead.`)
+  if (typeof meta !== 'string' && meta?.AffectedNodes === undefined) {
+    throw new TypeError(`Unable to parse the parameter given to getNFTokenID.
+      'meta' must be the metadata from an NFTokenMint transaction. Received ${JSON.stringify(
+        meta,
+      )} instead.`)
   }
+
+  const decodedMeta = ensureDecodedMeta(meta)
 
   /*
    * When a mint results in splitting an existing page,
@@ -46,7 +64,7 @@ export default function getNFTokenID(
    * if the PreviousFields contains NFTokens
    */
 
-  const affectedNodes = meta.AffectedNodes.filter((node) => {
+  const affectedNodes = decodedMeta.AffectedNodes.filter((node) => {
     if (isCreatedNode(node)) {
       return node.CreatedNode.LedgerEntryType === 'NFTokenPage'
     }